# 3x speedup: Hoist "allowed" check in OverlappingModel#propagate